+++ ricaun.DA4R.NUnit.Console/App.cs
@@ -8,5 +8,8 @@ namespace ricaun.DA4R.NUnit.Console
         public static string Name { get; set; } = "ricaun_DA4R_NUnit_Test";
         public static string Bundle { get; set; } = $".\\Resources\\ricaun.DA4R.NUnit.bundle.zip";
         public static string ForgeEnvironment { get; set; } = "release";
+        public static string ApsAccessToken => 
+            Environment.GetEnvironmentVariable("APS_ACCESS_TOKEN") ?? 
+            Environment.GetEnvironmentVariable("FORGE_ACCESS_TOKEN");
     }
 }